判定の破れと報告の読み取りの破れの器を新設し代表3コマンドを乗せる
判定層の失敗が541箇所の`Result<_, String>`へ散っていたため、破れの形を数え上げる2つの束を
検収の共通語彙へ置き、コマンド側がそれを使う向きにした。

判定の破れは9つの形(上限を超えた・下限に届かない・許容の範囲を外れた・一致するはずが食い違った・
零件のはずが残った・あるはずのものが1件も無い・有限でない・正でない・材料を数として読めない)を数え上げる。
何の判定かは`判定の名前`が、比べた数は`判定が比べる値`が持つ。比べそのものは名前のメソッドが行うため、
上限と下限を取り違えた形を呼び出し側で書けない。

報告の読み取りの破れは7つの形を数え上げる。読む側は`終了時報告`から`報告の行`と`報告の区画`を
型のまま受け取り、行の中の値は`報告の行の鍵`と`鍵に値が結ばれる形`で引く。
`報告の見出し`は実行時に組み立てた綴りも持てるようにした。

代表は3つを乗せ換えた。prop-drawが画素の判定、cluster-lightsが報告の行と区画の読み取りと決定性の判定、
auto-exposureが鍵と値の行の読み解きと許容つきの比べである。
画素の数どうしの比は`画素の占める割合`が持ち、分母の零をここで落とす。

+++ xtask/src/prop_draw/object_pixels.rs
@@ -15,7 +15,7 @@ pub(super) struct 囲む矩形 {
 pub(super) enum 物体画素の要約 {
     背景だけ,
     物体あり {
-        件数: u64,
+        件数: usize,
         矩形: 囲む矩形,
         明度最小: u8,
         明度最大: u8,
@@ -23,7 +23,7 @@ pub(super) enum 物体画素の要約 {
 }
 
 pub(super) fn 要約する(画像: &読み戻し画像, 背景色: [u8; 3]) -> 物体画素の要約 {
-    let mut 件数 = 0u64;
+    let mut 件数 = 0usize;
     let mut 矩形 = 囲む矩形 {
         x最小: 画像.幅().画素数(),
         x最大: 0,
